requirements.txt
matplotlib==3.7.5
numpy==1.23.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==10.4.0
pycocotools==2.0.7
torch==1.21.1
torch-fidelity==0.3.0
torchmetrics==0.5.0
torchvision==0.13.1
tqdm==4.67.1
